Day1_Streamlit_Basics/Task2_ChatApp_Groq_OpenAI/requirements.txt
streamlit==1.36.0
groq>=0.7.0
reportlab>=4.2.0
python-dotenv>=1.0.1
